(#18) Extracted static non-generic methods into a different module.
